+++ src/providers/recentmessages/Api.hpp
@@ -2,8 +2,10 @@
 
 #include <QString>
 
+#include <chrono>
 #include <functional>
 #include <memory>
+#include <optional>
 #include <vector>
 
 namespace chatterino {
@@ -28,8 +30,16 @@ using ErrorCallback = std::function<void()>;
  * @param channelPtr Weak pointer to Channel to use to build messages
  * @param onLoaded Callback taking the built messages as a const std::vector<MessagePtr> &
  * @param onError Callback called when the network request fails
+ * @param limit Maximum number of messages to query
+ * @param after Only return messages that were received after this timestamp; ignored if `std::nullopt`
+ * @param before Only return messages that were received before this timestamp; ignored if `std::nullopt`
+ * @param jitter Whether to delay the request by a small random duration
  */
-void load(const QString &channelName, std::weak_ptr<Channel> channelPtr,
-          ResultCallback onLoaded, ErrorCallback onError);
+void load(
+    const QString &channelName, std::weak_ptr<Channel> channelPtr,
+    ResultCallback onLoaded, ErrorCallback onError, int limit,
+    std::optional<std::chrono::time_point<std::chrono::system_clock>> after,
+    std::optional<std::chrono::time_point<std::chrono::system_clock>> before,
+    bool jitter);
 
 }  // namespace chatterino::recentmessages