+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 - Dev: Move `clang-tidy` checker to its own CI job. (#4996)
 - Dev: Refactored the Image Uploader feature. (#4971)
 - Dev: Fixed deadlock and use-after-free in tests. (#4981)
+- Dev: Load less message history upon reconnects. (#5001)
 
 ## 2.4.6